Move iterative classes into leoAst.py and test_leoAst.py
